[FIX] account/saas~16.5: complexity of matching number computation
The recursive query was done in O(n^2) because it was going through the
whole graph of matching, starting from each node.
This new approach is only going through each link once, by using a
temporary table to remember each subgraph and be able to find them with
an index, speeding up considerably databases containing a lot of complex
reconciliation of a big size.

For the upgrade of a database with
* 13M account_move_line
*  3M account_partial_reconcile
* one complex matching of 17k lines and multiple others of ~1k lines

The time of this part of the upgrade went from more than 38 minutes to
4s.

closes odoo/upgrade#5537

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>