[FIX] base/0.0.0: handle forward-port of payment modules
The deletion of payment modules requires a cleaning of the related
providers.
In order to dynamically support the forward-ports, new modules are
actually ignored during upgrade by removing them as if they were
deleted.

To allow a clean suppression of payment modules, we must cheat and
remove the blocking foreign keys before deletion.
Those FK will be readded by the ORM during the upgrade of the payment
module.

closes odoo/upgrade#5539

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>