[IMP] project/saas~17.2: remove `closed_task_count` field in project
Before this commit, the `closed_task_count` has been almost deprecated
since version 17.0 because in the project kanban view, the number of open tasks
and so to do that a soustraction was made between the total of tasks in the
project and the number of closed tasks in that project. However, the result
cannot be formatted since the format function is not available in the kanban
record. To solve that issue and avoid having `NaN` when the soustraction is
made between 2 formatted numbers, a new computed field has been introduce to
display the number of open tasks in the project.

This commit removes the `closed_task_count` because that field has been
completely deprecated in 17.2 to use `open_task_count` instead.

Related PR: odoo/odoo#150326

closes odoo/upgrade#5566

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>