[FIX] sale/saas~16.5.1.2: linking attachments to product.document
- after the pr: https://github.com/odoo/odoo/pull/132739/files
Introduce new model named 'product.document' to hold documents
linked to a given product template/variant.
also added new smart button on there to show the linked documents.

-Basically, In version we don't have smart button of "Document"
and it's introduced in version 17.0.
The thing is that smart button is working fine when we upload a document
after migration and it is considering only those documents which are
uploaded after migration.So, the problem here is if we have documents
for a product in version 16.0 and after migrating those documents
should be visible in the smart button and the count as well but they
are not visible because no records available in the product_document model.
For handling this particular scenario we are inserting/attaching the records
of documents from ir_attachment model into product_document model
so that it can show the exact count and display documents in the
smart button for all the products correspondingly.

- Steps to reproduce:
1. create database with 16.0
2. attach any document with product template/variant
3. upgrade to 17.0
4. check the form view of that product template/variant
Issue: smart button named 'Documents' with count 0

- Before fix:
================
smart button comes with 0 count even the documents are there in attachment.

```sql
syna_1203926_17.0=> select * from product_document;
 id | ir_attachment_id | create_uid | write_uid | active | create_date | write_date | attached_on
----+------------------+------------+-----------+--------+-------------+------------+-------------
(0 rows)
```

- After Fix:
==============
Documents from attachment shows with appropriate records

```sql
syna_1203926_17.0=> select count(*) from product_document;
 count
-------
   145
(1 row)

```

-OPW-3672233
-UPG-1203926

closes odoo/upgrade#5571

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>