[FIX] base/15.0.1.3: Recursive administration categories
In 15.0, a check has now been introduced to prevent the existence
of recursive categories. But by using the noupdate = True flag it
was still possible for a customer to have a recursive one on the
`module_category_administration` and
`module_category_administration_administration` category

If any cycle exists in the categories we break them.

upg-1221877
upg-1221716

closes odoo/upgrade#5574

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>