[IMP] point_of_sale/saas~17.3: remove old customer display fields
In the corresponding commnity pr, the customer display feature of
pos is completely refactored and also moved to a separate new module:
`pos_customer_display`.

In this commit we update the `point_of_sale` and `pos_iot` modules
to reflect the new changes.

Some fields are simply renamed. In other cases the modelling is changed.

https://github.com/odoo/odoo/pull/148870
https://github.com/odoo/enterprise/pull/54501

closes odoo/upgrade#5585

Related: odoo/odoo#148870
Related: odoo/enterprise#54501
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>