[FIX] account_reports/16.0: Conditionally set `cross_report`
opw-3591708 upg-1173517

Setting `cross_report` on all expressions is wrong and causes reports to
misbehave. In fact, due to a known limitation of the reporting engine,
when an expression tries to make a cross report aggregation including
another expression of the **same** report, with a different `date_scope`,
the cross referenced expression will "force" its own `date_scope`,
making the computation wrong.

closes odoo/upgrade#5612

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>