[IMP] mail/saas~17.2: allow to configure activity templates deadline
We rename the field plan_date_deadline to plan_date of the wizard
mail_activity_schedule because that date doesn't represent the deadline but
the base date of the plan.

We remove the computed field assignation_summary of mail.activity.plan because
we now compute it in the wizard as the summary now includes the deadlines of
each activity which depends on the plan date defined in the wizard.

We rename the computed field plan_assignation_summary of mail.activity.schedule
to plan_summary as it is no longer about who will be assigned to the activity.

Note that we don't upgrade the plan template delays. Before, the deadline of
the activities created when scheduling an activity plan were determined by the
plan date + the delay configured in the related activity type. So to keep the
plan unchanged, we should set, in this upgrade, for each template, the delay
configured on their related activity type (delay_count and delay_unit) and
'after_schedule_date' for the delay_from. We don't do it as it was not logical
in the previous version to use the delay of the activity type for that. As a
consequence, the existing plan will have the default values which are a delay
of 0. But user will be able to see this change each time they schedule an
existing plan thanks to the summary that is displayed at that moment.

See odoo/odoo#152215
See odoo/enterprise#55597

Task-3639909

closes odoo/upgrade#5619

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>