[IMP] event/saas~17.4: remove redundant fields
notification_type and template_model_id
were used in conjunction with a reference field (template_ref) to select
what kind of template we use.

As the reference already defines the model, and the model implies
the type of notification. And because we don't care about those
when no template is set. It doesn't make sense to keep those fields around.

In terms of UI, instead of picking a notification type from a selection
then setting the model_id and using that
as the reference for template_ref
We just set a selection on template_ref and make use of the reference field's
ability to let people pick from a selection field in the front-end.

odoo/odoo#152424
odoo/enterprise#60625

task-3256524

closes odoo/upgrade#5914

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>