[FIX] account/saas~13.4: fix slow query in payment-refactoring
upg-1236571

This DB hs been processed in the platform several times. Sometimes it was
processed in ~1h, sometimes it took ~11h. The reason was pinned down to this
query:
https://github.com/odoo/upgrade/blob/a934c595298c760e59cb3b7ac6e019a5be7bac42/migrations/account/saas~13.4.1.1/pre-10-payment-refactoring.py#L574
and different planner decisions. In the "good" runs (< 1min), the execution
plan was mainly made out of hash joins. In the "bad" runs (> 11h), the plan
mainly used nested loops. The latter are slow, because this causes a lot of seq
scans over `account_payment_pre_backup`. Instead of trying always nudge the
planner in direction of the hash joins, support the nested loop plan by adding
indexes on all `account_payment_pre_backup` columns that are involved in any
join condition in any of the "payment_refactoring" scripts. Experiments show
that with these indexes, both kind of plans run in ~35s.

closes odoo/upgrade#5630

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>