[IMP] l10n_in/saas~16.4: remove view
With the initial bugfix of odoo/odoo@1146df50a in 15.0, we had to
introduce 2 views to avoid l10_in invoice report affecting that
same report for other countries.

But with the invoice report changes in odoo/odoo@fc0464427, the
`account.report_invoice_with_payments` now use `account.report_invoice`
so we have to remove one of the added view for databases that were
created or which updated `l10n_in` with that intial fix.

closes odoo/upgrade#5646

Task-id: 2504287
Related: odoo/odoo#152848
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>