[IMP] base/0.0.0: disable failing QWeb reports
QWeb views are currently skipped during view validation because of the empty
`model` field.
This commit adds the steps to validate the combined arch of qweb reports.

Context:
- Prior to 16.3 (inclusive)
  - Custom reports from studio would be created by copying the arch of an
standard report into a new view
  - Modifications on reports (from standard, custom code or studio) would be
done by applying a list a xpaths entries, encoded in a view file
- After 16.3
  - Same behaviour for custom reports as before
  - Modifications on reports are now encoded by copying the whole modified arch
into a new studio view and applying the changes in the copied view itself

That said, this commit targets mainly the combined arch strategy of 16.3 and
older versions, since the new strategy doesn't rely much on xpaths.

Note that this check is useless for primary standard reports per se, since these
will get reloaded when loading the `.xml` related to it during the upgrade
(provided that `noupdate=0`).
However, the combined arch (of standard or custom reports), after applying the
studio modifications will benefit from this check, given that these
modifications codified as xpaths are prone to be turn obsolete after the upgrade

For later commits,
- apply heuristic fixes on qweb views
- check for errors on COW views
- check for attribute errors on reports

Part-of: odoo/upgrade#5995
Co-authored-by: Paulo Medeiros <palm@odoo.com>
Co-authored-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>