[FIX] hr_payroll/saas~16.5 : update salary rules as standard
Due to a generic PR merge, which can be found at
[https://github.com/odoo/upgrade/blob/master/migrations/hr_payroll/saas~16.5.1.0/post-migrate.py#L27],
the values of the amount_python_compute field are displaying
(.get) in both concatenated categories.Additionally, because of this,
if one of its values is null, it returns a NoneType value,
leading to a type error stating that a float cannot be added with NoneType.
Therefore, if we replace it according to our standard, which can be found here
[https://github.com/odoo/enterprise/blob/17.0/hr_payroll/data/hr_payroll_data.xml#L71],
then it will be able to compute the field (amount_python_compute).

closes odoo/upgrade#5680

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>