[FIX] l10n_din5008_industry_fsm/saas~16.3: remove a field on task
On a FW-port the field `l10n_din5008_document_title` was added
on model `project.task` on 16.2 (and previous versions)
but not on 16.3.

This happened since a new title was added in 16.3 to the
"Field Service Report" (previously "Task Report").
So the DIN5008 title was dropped on 16.3.

This commit removes the aforementioned field (on upgrade).

see odoo/enterprise#45625 (16.3)

closes odoo/upgrade#5684

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>