[FIX] analytic/saas~16.5: avoid merging plans from the same company
The merging of analytic plans with the same name is supposed
to be for the same plans that are duplicated across companies
but we have the possibilities of using the same name for
multiple plans under the same company but using a different
hierarchy. In this case we do not want those to be merged.

closes odoo/upgrade#5689

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>