[FIX] website/saas~17.2: update s_embed_code snippet
The embed code snippet was modified with [1] by adding a `template`
element, so as to not launch `script` tags in edit mode. In stable
versions, the template is added when the user edits their page. On the
contrary, the code in master (see [master PR]) considers by default that
there is a `<template>` inside the snippet.
This upgrades existing snippets which wouldn't have a `<template>` yet.

[1]: https://github.com/odoo/odoo/commit/26d8fb39e193fc2652c7f420797f559b460ac2d3
[master PR]: https://github.com/odoo/odoo/pull/153928

closes odoo/upgrade#5693

Related: odoo/odoo#153928
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>