[FIX] l10n_lt_reports/saas~17.3: remove obsolete records
In the related enterprise PR (see below) the reports were updated to
use tags instead of codes.
As part of this the profit and loss information in the BS was
updated to refer directly to the P&L (cross_report).
In the process some account.report.expression records were removed.

task-3701151

see odoo/odoo#154407
see odoo/enterprise#56825

closes odoo/upgrade#5706

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>