[FIX] base/saas~16.5: create forgotten columns for server actions
**Before this commit**
Since [1], when the base module upgrades and these two columns
are created ('update_path' and 'update_related_model_id'), it triggers
the '_compute_crud_relations' compute method for some fields.
As this overrides some server actions values (i.e. 'crud_model_id'),
this is not what we want and we should avoid to recompute.

**After this commit**
The columns are created in a `pre-` upgrade script (before the ORM is loaded).
That way the base module upgrade will not create the columns and thus the
compute method will not get called.

[1]: odoo/odoo@253c214b5b76a13f2cc

Part-of: odoo/upgrade#5677