[FIX] l10n_mx/saas~16.2: Fix ValueError because of missing xmlid
While pre-loading the data through the chart template, if the company
has a parent company, part of the data is removed:
(https://github.com/odoo/odoo/blob/17.0/addons/account/models/chart_template.py#L182-L185)

This way, if the corresponding missing xmlid `account.{company.id}_{xmlid}`
is missing, it will not be created and a `ValueError` will raise when
the xmlid is referenced.

By sorting the company array with the root companies first, we avoid the issue.

closes odoo/upgrade#5708

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>