[FIX] analytic/saas~16.5: don't create x_plan{id}_id for child plan
The new logic for analytic lines is to create a custom field for each
root analytic plan. But currently, the migration creates this field also
for child plans.
The fix is to not create a column for child plans and also to link the
analytic account in its root plan column.

opw-3741514

closes odoo/upgrade#5715

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>