[FIX] appointment/saas~16.5: fill appointment_booker_id field
Introduced in odoo/enterprise#45332, a new feature allows
the people booking an appointment to give a list of
emails to add guests to the appointment. A new field
appointment_booker_id of type res.partner is introduced
to explicitely set the partner booking the appointment
in the case of the appointment flow.

It was not set in the upgrade script. Correct this for
future events, by setting the booker for events linked
to an appointment type, and using, in this order:

1. the partner linked to the event create_uid if in
the meeting partners
2. the first partner among the meeting partners that
are not the one of the responsible, if any
3. the partner of the meeting responsible

A test is included.

Task-3701005

closes odoo/upgrade#5734

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>