[FIX] sales_team/0.0.0: ensure right default value
`crm` has `True` as default value for `crm.team.use_opportunities`.
https://github.com/odoo/odoo/blob/c7f982774654459a2db6decd6ae1948e05e7d622/addons/crm/models/crm_team.py#L26

`crm.team` model is defined first in `sales_team` and some teams are
already crated there, in particular the `Sales team:
https://github.com/odoo/odoo/blob/c7f982774654459a2db6decd6ae1948e05e7d622/addons/sales_team/data/crm_team_data.xml#L4-L8

The issue is that when those teams are created during the upgrade, the
ORM has yet to load `crm` module therefore it creates the teams with
`NULL` value in `use_opportunities` -- for the ORM `NULL` is false
always. Thus when the `crm` module finally loads the field the ORM
assumes the NULL value is False and never sets it to True. This is
something unfixable without deep changes in how the ORM handles boolean
values. Therefore current patch ensures that NULL values are set to True
when `crm` is already installed in the DB, to have a consistent default
value for `crm.team` records.

upg-1389119
tgb-1091

closes odoo/upgrade#5739

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>