[IMP] {base,sale}/saas~17.3: merge *_product_configurators module
Since ee3bf99, `sale_product_configurator` is auto-installed. The
settings to enable or disable it have been removed and the only way to
deactivate the product configurator is to uninstall the module but then
it provides a poor UX (eg. in `website_sale` buy now button on /shop
page adds the first available combination to the cart instead of opening
the configurator).

Therefore, we have decided to merge modules to improve the readability
and reduce the complexity of our codebase.

See odoo/odoo#155449
See odoo/enterprise#57577

task-3763473

closes odoo/upgrade#5750

Signed-off-by: Victor Feyens (vfe) <vfe@odoo.com>