[ADD] documents/17.0 don't delete archived documents immediately
Since odoo/enterprise@2254eb07fd42ec88341d9d7ac92f09a379c6fa55, archived
documents are deleted after 30 days and placed in a trash bin.
As the delay is base on the `write_date`, documents archived for more
that one month before the upgrade will be removed as soon as the
auto-vacuum is run.
Avoid this by forcing the write_date of those documents.

opw-3751629

closes odoo/upgrade#5760

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>