[FIX] base/tests: menu crawler test: skip know failing menus
Some menu are known to fail on CI. Skip them explicitly.

In the case of `l10n_fr_hr_holidays`, the failure is due to the fact
that the reference time off type should be configurated on the
french companies[^1][^2].
As matt force the main company to be in the country of the installed
`l10n_` module, this verification fail. Skip the menu in this case.

[^1]: https://github.com/odoo/odoo/blob/3b018cfe7e905515bec16c6023519e738a831a63/addons/l10n_fr_hr_holidays/models/hr_leave.py#L26-L27
[^2]: https://github.com/odoo/odoo/blob/0dfee1606b3de0815989d1e1d9a037f66b47daff/addons/l10n_fr_hr_holidays/models/res_company.py#L15-L19

closes odoo/upgrade#5766

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>