[FIX] l10n_mx_edi/saas~16.5: always recompute attachment_origin
We need to always recompute "attachment_origin" when upgrading
to 17.0 (aka 16.5). Previously when the query was in pre, it had
to be under the condition that the table exists (it only exists
from 16.4 onward). This resulted in dbs originating from before
16.4 (16.0 for example) not having their edi docs recomputed at
all, because no documents were marked for recomputation for the
cron that runs later, after the db is restored.

closes odoo/upgrade#5774

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>