[IMP] l10n_au_hr_payroll/saas~17.3.1.0: Termination payment
The termination payments are added as a separate
salary structure. That only allows termination payment
to be made through a new payslip after the final payslip.

This commit allows the rules from the termination structure
to be added directly to the last payslip along with the default
structure.

see https://github.com/odoo/enterprise/pull/57884

task # 3639270

closes odoo/upgrade#5806

Related: odoo/enterprise#57884
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>