[FIX] account/saas~16.2: Only explode queries on tables with id
During the Accounting migration to saas~16.2, we are changing property fields
of account.tax.group into regular fields, by making copies of account_tax_group
entries with the new company_id field depending on which companies are linked to
which tax groups. We detect that by looking for tables that have a reference to
both res_company and account_tax_group, and afterwards update the tables to
point to the copies.

However, since we're doing that in parallel with explode_execute, the tables
need to have an `id` which is not always the case, e.g. when there's an m2m
res_company-account_tax_group relation (as happens with the linked upg's
customer data).

We add a check for the `id` field so that in its absence the query can be run
non-parallel.

upg-1381715
Ticket #3736698

closes odoo/upgrade#5789

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>