[FIX] sale_subscription/saas~15.3.1.1: Restrict the duplicate record
In the v16 standard, we cannot insert duplicate iteration and pricing
records in product_pricing.
When trying to enter the update, this error occurs
'You cannot have multiple pricing for the same variant,
recurrence, and pricelist'
In our script we insert data through unique rows
we don't select a particular column
so data with different price are also inserted into the table.
Which behavior is not as per v16.

closes odoo/upgrade#5799

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>