[IMP] website/0.0.0: neutralize COWed core layout templates
This commit removes COWed versions of core layout templates such as:
- `web.layout`
- `web.frontend_layout`
- `portal.frontend_layout`
- `website.layout`
- `website.navbar_toggler`

The `website.navbar_toggler` is reset because it should not contain any
arch diff, but before 17.0 it was possible to mess with this view by
clicking on the menu toggler and typing text or removing the icon or
stuff like that. In 17, it was marked as not editable.
For those reason, we reset that view content too.

task-3753159
opw-3730155
opw-3735050

closes odoo/upgrade#5801

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>