[FIX] account/saas~16.3: disconnect wrong inactive taxes
When the charts of accounts are loaded, all standard taxes
including the inactive ones are updated.
The tax update includes a full recreation of the tax if it
was modified, and if it was not then only updates the tags
on repartition lines.
The update of tags triggers the validation of the lines
which can fail for some taxes.
This script renames the inactive taxes that will fail
validation during the upgrade and allows the CoA reload
to recreate them properly.

closes odoo/upgrade#5763

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>