[IMP] mrp_workorder/saas~17.{1,2}: better filename
It's more about the symlink filename.
Using the classic `pre-migrate.py` for a symlink lead to unexpected
changes to the `saas~17.1` script made by developers unaware of symlink
trick that lawfully edit the `saas~17.2/pre-migrate.py` file without
seeing it actually change the `saas~17.1` script.

And this is not a theory. It actually happens at
https://github.com/odoo/upgrade/pull/5780#discussion_r1521628784

Oversight of https://github.com/odoo/upgrade/pull/5777

closes odoo/upgrade#5811

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>