[FIX] l10n_ph/17.0: Fix taxes' country with correct company
A case not considered by https://github.com/odoo/upgrade/pull/5797/
Some databases have fixed the country for the company but not for the taxes,
so the error keeps on happening as the fix is not applied in that case.
Since the tax groups are already filtered by xmlid, it should be safe.

upg-1415087

closes odoo/upgrade#5815

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>