[ADD] portal/saas~17.3: remove portal_record_layout template
This commit deletes the `portal.portal_record_layout` as it is not needed
anymore due to a design choice to remove the card design in the new
customer portal redesign.

Prior to this commit, this template was used to apply Bootstrap classes
related to the `card` component such as `card-header`, `card-body` if
a `card-x` variable was set to true.

In Odoo 17 we already took a first step and remove some occurrences of
the template but it was still used by some views.

As we now removed the card design from all the views of the customer
portal, this template is not needed anymore and can be remove.

See odoo/odoo#154424 & odoo/enterprise#56830

closes odoo/upgrade#5814

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>