[IMP] base/0.0.0: add warning for path modification in ir.actions
[1] introduces a path-based routing in the webclient, to do this, a path
can be added to the actions. The path needs to be unique, and stay the
same all the time. This requirement is to avoid breaking existing URLs
(for instance, URLs that has been shared or bookmarked).

This commit, add a warning when upgrading, and remove the new path
value, to leave the old one.

[1]: https://github.com/odoo/odoo/commit/c63d14a0485a553b74a8457aee158384e9ae6d3f

closes odoo/upgrade#5816

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>