[FIX] project/saas~16.3: keep hidden subtasks hidden
Prior to this commit, we just removed the `display_project_id`
field, and because of that, we lost track of the subtask that
we wanted to display in the project or not.
In next version, the field `project_id` is the one to hold this
information. So, by setting it to null when `display_project_id`
was null, we make sure to keep track of the subtasks we didn't
want to display in the project.

The reason why we didn't do that at the time, is that `project_id`
was the only "project" field on a task. Yet by unsetting its project,
a task would have lost it timesheets and so it would have constituted
a regression. However, in saas-16.3, it was ultimately decided to add
a computed non-stored project field `project_root_id`, that would allow
subtasks with no project (aka not visible in the project) to have
timesheets. That also explains why we can now unset `project_id` on
those.

closes odoo/upgrade#5818

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>