[FIX] spreadsheet_edition/17.0: copy image attachment
Steps to reproduce:

- create a new spreadsheet
- insert an image
- duplicate the spreadsheet
- delete the original spreadsheet (deleted, not archived)

When duplicating a spreadsheet, the images src `/web/image/<attachment_id>`
are still pointing to the image attachment of the original spreadsheet.
If the original is deleted, the attachment is gone and the image disapears
from the remaining spreadsheet.

In prior versions, documents were archived and never deleted (unless
explicitely deleted from the trash). Since 17.0 however, documents in the
trash are automatically deleted after 30 days.

This commit ensures every image attachment is linked to the correct
spreadsheet record. If it's not, the attachment is copied and linked to
the record.

Follow up PRs are coming for master.

odoo/enterprise#58295
Task: 3775364
opw-3751629

closes odoo/upgrade#5795

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>