[IMP] account_asset/saas~17.4: compute newly stored fields
When handling account moves involving assets, it can be
challenging to determine the relationship of the move to
the asset. This could be a sale, disposal, purchase, or
positive/negative revaluation. Previously, this information
wasn't stored directly, but it is now saved in the new field
asset_move_type.

However, asset_move_type needs to be computed for existing
records. The query to compute this field differs from the
compute method used during the regular creation/modification
of a move, primarily because values of this field can be set
in quite widespread areas of the code, such as the
account.modify wizard, the account.asset model, or the
account.move model.

For existing records, this upgrade script sets the correct
value based on simple heuristics, such as the number of
account move lines or the asset_value_change field value.
These values might have been altered by the user for any
reason, which might lead to incorrect data. To solve this,
the user can manually modify the value of the asset_move_type
in the "Other Info" tab of account moves.

Additionally, the net_gain_on_sale field also needs to be
computed for existing records for similar reasons.
This field is set on the asset and defines the net value
obtained when either selling or disposing of an asset.

See https://github.com/odoo/enterprise/pull/57960
task-3615633

closes odoo/upgrade#5880

Related: odoo/enterprise#57960
Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>