[IMP] spreadsheet_edition/saas~17.3: change pivot measures and groups
This commit changes the way pivot measures and groups are stored in the
ADD_PIVOT command. This is done to be able to add properties to measures
and groups in the future. (e.g. the measure aggregate or the order of a group)

Before this commit:
```js
{
  type: "ADD_PIVOT",
  pivotId: 1,
  pivot: {
    colGroupBys: ["A"],
    rowGroupBys: ["B"],
    measures: ["C"]
    ...
  }
}
```

After this commit:
```js
{
  type: "ADD_PIVOT",
  pivotId: 1,
  pivot: {
    colGroupBys: [{ name: "A" }],
    rowGroupBys: [{ name: "B" }],
    measures: [{ name: "C" }]
    ...
  }
}
```

See https://github.com/odoo/odoo/pull/158253
and https://github.com/odoo/enterprise/pull/58957

Task-id 3813563

closes odoo/upgrade#5834

Related: odoo/odoo#158253
Related: odoo/enterprise#58957
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>
Co-authored-by: Lucas Lefevre <lul@odoo.com>
Co-authored-by: Pierre Rousseau <pro@odoo.com>