[FIX] hr_skills_survey/saas~17.3: fix certification resume line end_date
When completing a certification (survey), a hr_resume_line was inserted in the
database with a start_date and an end_date set to the completion date, causing
the certification to be expired right away. After correcting the code, we
update those existing resume lines with the end_date set to NULL and
expiration_status set to "valid" (as it is a computed stored field based on
end_date), meaning they will never expire.

Also, when completing a certification with an empty description, a
hr_resume_line was inserted with a description set to the string "False". We
then also update each translation that is set to the string "False" to the
empty string.

See odoo/odoo#142333

Task-3389395

closes odoo/upgrade#5841

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>