[FIX] l10n_mx/16.0: create missing group
During upgrade of version 15.0 to 16.0,
for mexican localization, account.group has
not been created from account.grouptemplate
and because of that group_id was missing
(not linked) with account_account table.

when same database upgrade from 15.0 to 17.0
will get account.group and also related
group_id linked with correct account_account.
because for version 17.0

we call `try_loading` from generic script
in saas~16.2, so it wll create missing group
and link it with account which was not happen
with version 16.0.

when We connect fresh database and installed
l10n_mx, on that also group id connected with
account only when we configure chart of template
from configuration setting.

so to create group from account.group.template
we need to call this method.
https://github.com/odoo/odoo/blob/55c605f44d3a4f280946f1c1fec3f5e1534b4ef9/addons/account/models/chart_template.py#L855C14-L855C37

Unicorn-task:
https://www.odoo.com/web?debug=1#id=3729675&cids=21&menu_id=4720&action=333&active_id=6443&model=project.task&view_type=form

upg-1312543
upg-1308309

closes odoo/upgrade#5846

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>