[IMP] hr, resource/saas~17.5: Populate calendar_id for flexible employee
This commit aims to add fields related to flexible working hours, and
moves the existing field 'full_time_required_hours' of resource_calendar
defined in hr_payroll, into community's resource.

1) Flexible Hours related fields
In order to introduce flexible working calendar, all employees with a
flexible working schedule (defined as 'calendar_id = None') need a
calendar assigned with field 'flexible_hours' set to True. This commit
aims to populate the field 'calendar_id' by creating a flexible calendar
based off the default calendar of the company.

Additionally we create the fields 'is_flexible' and 'is_fully_flexible'
in the 'hr.employee' model, and populate the 'is_flexible' field based
on the field 'flexible_hours' created before in the 'resource.calendar'.

Note that the field 'is_fully_flexible' is not populated, since it is a
new category of resource calendar.

2) Move full_time_required_hours field
By moving the full_time_required_hours field to the resource
model, we can now have the denominator of the ratio to calculate the
progress bar for flexible resources in the planning gantt view, when
hr_payroll is not installed.

closes odoo/upgrade#5851

Task: 3762895
Related: odoo/enterprise#58599
Related: odoo/odoo#157577
Signed-off-by: Yannick Tivisse (yti) <yti@odoo.com>