[ADD] base/0.0.0: nullify res_users.action_id when action is removed
Field `action_id` on `res.users` is used to set the default action
to be executed upon logging in. Its corresponding column however,
is not a typical one in the sense that it doesn't have a foreign
key constraint linking it to `ir.actions`. So when an action
is removed, the column wouldn't be automatically nullified.

We rectify this issue by running an additional check with this
upgrade script. If the column is referencing an action that no
longer exists, it gets wiped.

Field `action_id` on `res.users` exists since saas~11.3.

closes odoo/upgrade#5869

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>