[FIX] sale_stock_renting/saas~17.2: Adapt old rental rules
Due to the new way push rule work, it was not viable anymore to keep the
old rental rules as 'push'. Instead, we decided to use pull rules
instead to pull from the Rental location, then follow the new push flow.

This meant that we need to change the existing rental stock.rules for
users already using this feature to convert them to the new ones.

See odoo/enterprise#58758

Part-of: odoo/upgrade#5873