[IMP] industry_fsm_report/saas~17.3: rename view of worksheet formview
In this commit, the form view dialog view has renamed in order to simplify
the understanding of the inherited worksheet template view.
This view was previously used to hide the studio button, but is no longer
required since it used to display the studio button for the design templates.

Related Community PR: odoo/odoo#152431
Related Enterprise PR: odoo/enterprise#51749

task-3504607

closes odoo/upgrade#5649

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>