[FIX] account/saas~17.3: make preceding subtotal label translatable
Tax groups can have a "preceding subtotal" label which is put on the
invoice to indicate the amount before taxes from this group are applied.

The label was not translatable before, so users with multiple languages
could only define it in one. This commit fixes that by making the field
translatable.

task-3730774

See https://github.com/odoo/odoo/pull/160466

closes odoo/upgrade#5892

Related: odoo/odoo#160466
Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>