[FIX] sale_timesheet/saas~17.1.1.0: fix standard view translation
While opening project app in Version saas~17.1 with other language
such as "French / Français" and "German / Deutsch" except 'en_US',
it will raise error mentioned below. Because in V17.1 the span tag
has been added in 'sale_timesheet.project_project_view_form' view.
Hence, during upgrade, the translation of it is not been updated
as per standard because we update translation of upgraded version's
records from previous version's translation directory that's why it
will not update.

So, This PR will update the arch of the view manually for this
languages.

This is only necessary when the industry_fsm_sale is installed, as
then only this view will be loaded which is the root cause of the
error.
See enterprise/saas-17.1/industry_fsm_sale/views/project_project_views.xml#L56

opw-3780242,3790704
upg-1446734,1448049

closes odoo/upgrade#5898

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>