[IMP] survey/saas~17.3: remove unnecessary template
The xml template for the summary stats was previously called in
each relevant question type template using a 't-call'.
However, due to a ui improvement of the results page, the summary stats
need to be displayed alongside the other general question info.
The use of a template is thus not needed anymore. Its content is moved
inside the general question info template and its content visibility is
restricted using a 't-if'.

see: odoo/odoo#154904

Task-3707687

closes odoo/upgrade#5901

Signed-off-by: Christophe Simonis (chs) <chs@odoo.com>