[IMP] website/saas~17.4: keep previous behavior for cookies
Upgrade script for [PR-144624], which improves the handling of third-
party cookies on website.
By default, if the cookies bar is active, the new feature blocks third-
party services that track users when they are embedded (like YouTube)
and shows a placeholder message instead of the iframe until the user has
given their consent.
While this is the right way to block trackers and respect the user's
consent, it is a big change from previous versions. We thus revert the
default for upgrades.
Only exception: if a client uses Google Analytics and activated the
cookies bar, they expect Analytics to be blocked (which was already the
case before this task). In that case, we keep the default.

[PR-144624]: https://github.com/odoo/odoo/pull/144624

task-3027209

closes odoo/upgrade#5905

Related: odoo/odoo#144624
Signed-off-by: Quentin Smetz (qsm) <qsm@odoo.com>