[FIX] hr_holidays/saas~16.5: Set correct cap_accrued_time
Before 17.0, if you wanted accrued time from an accrual plan level to be
unlimited, you would set "maximum_leave" to 0. In 17.0, there is now a boolean
"cap_accrued_time" that determines whether the accrued time is limited or
unlimited.

When the database is upgraded, "cap_accrued_time" is set to true and the
"maximum_leave" is set to 0. This causes issues with the new logic making some
accrual plans to lose allocated days.

Unicorn task: 3836753

closes odoo/upgrade#5913

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>