[FIX] account/17.0,saas~17.2,saas~17.3: rm currency col on payment term
See odoo/odoo#161767

The currency_id field on the account.payment.term is exclusively used to
display the relevant currency on the live preview of the payment terms.

There is no substantive reason for it to be stored and, in fact,
computing it from the env.company is preferable since it will display a
more appropriate currency in the preview.

This was the justfication for making this field non-stored.
This targets both version 17.0 and saas~17.2 since the change was made
in stable.

closes odoo/upgrade#5919

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>