[FIX] stock/saas~17.3: set product_category_name as store=False
Product_category_name is set as `store=True` only for frontend purposes.
Since 17.2 we can now group by a related non-stored field if it's
compute_sudo=True (which is the default for related field). Marking
this field as `store=False` saves lot of time when changing
the parent_id of a product_category.

See odoo/odoo#162085

closes odoo/upgrade#5928

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>