[FIX] sale_subscription/saas~15.3: Fix Missing sale order line
Last week this [fix](https://github.com/odoo/upgrade/pull/5799)
got merged due to this sale order line are not inserting the
reason is below and how I am correcting.

-First, Why product_pricing duplicate records were going to
insert because their are subscription_line with different unit price
other then the price menitoned on the product. This valid client
can sale product at his choice price and this is editable their.
due to this while group by and distinct duplicate records are going to
select with different price and later on for the particular
sale_order_line ``pricing_id`` is going to set with join. Earlier
it is working fine because records are coming duplicate and the id
are also their after this [fix](https://github.com/odoo/upgrade/pull/5799)
only single record is coming due to that most of sale_order_line
are not inserted and after doing left join it will insert pricing_id
for those sale_subcription_line it created the product_pricing for
left we will set null.

upg-1517377

closes odoo/upgrade#5946

Signed-off-by: Alvaro Fuentes Suarez (afu) <afu@odoo.com>